Improve accuracy of profiler overhead tracking
Before this commit, the profiler tracked its overhead using a
tick := time.Now() at the profiler method start and time.Since(tick) statement
at the method's end. This approach however fails to account for the actual time
spent invoking the profiler hook (stack setup, register push on
entry/pop on exit) or the time spent invoking time.Now() and
time.Since().

The actual overhead is in the nanosecond range but causes a significant
skew in the total time measurements if the profiled function is invoked
a large number of times.

This commit attempts to improve the tracking of overhead by using a
calibration loop to estimate the mean time required for time.Now/Since
and runtime overhead for invoking profiler methods and factoring that
time in overhead calculations.

Since we are using the mean time, our calculations are still not 100%
accurate but allow us to reduce our skew to ~40ms for a method invoked
1M times (~40ns untracked overhead per call).